requirements.txt
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==5.10.0
gspread-asyncio==1.9.0
h11==0.14.0
idna==3.6
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
pandas==2.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wsproto==1.2.0
